Implement PIN block encryption for AES DUKPT
Note that this applies to encoded PIN blocks, not raw PINs, and also
requires the encoded PAN block for the encryption of ISO 9564-1:2017
format 4.

Unfortunately the Supplement to ANSI X9.24-3:2017 -- Test Vectors
document only provides encrypted PIN block data for the first eight
transactions of AES-128 BDK. Only the PIN keys are available for the
other test cases, while the intention of this library is to encapsulate
the derivation of working keys and their use. Therefore, to verify the
PIN encryption for AES-256 BDK, the encrypted PIN blocks are decrypted
using the PIN keys provided by the test vectors to confirm that the
original PIN block can be recovered.