Implement CMAC generation and verification for AES DUKPT
AES DUKPT allows for the derivation of AES keys to be used for message
authentication but does not specify the exact technique. CMAC seems to
be a popular choice in the card payment industry for AES DUKPT. By
contrast, CBC-MAC has many pitfalls.

Unfortunately X9.24-3:2017 does not provide test vectors for the
response MAC key.