Implement HMAC-SHA256 generation and verification for AES DUKPT
AES DUKPT allows for the derivation of HMAC keys to be used for message
authentication but does not specify the digest to be used with HMAC.
SHA1 is increasingly unpopular in PCI standards, making SHA256 the
likely next popular choice. PCI hasn't embraced SHA3 yet.

Unfortunately X9.24-3:2017 does not provide test vectors for the HMAC
keys.