Implement PIN encryption/decryption for AES DUKPT
Note that this implies encryption/decryption of raw PINs and builds on
the existing implemention of encryption/decryption for encoded PIN
blocks. It is the responsibility of the caller to ensure that the PAN is
already in compressed numeric format (EMV format "cn"). This is the same
format as EMV field 5A which typically contains the application PAN.