Let dukpt-ui require dukpt libraries
The dukpt libraries should either be provided by the parent project (as
will be the case for the default build) or by the platform (when
building dukpt-ui as a standalone project).

It is also useful to alias the dukpt_tdes and dukpt_aes targets such
that they can be referenced via their dukpt:: namespace regardless of
whether the parent project provides these targets or whether the
installed CMake config file provides these targets.