Implement PIN encrypt/decrypt helper functions for dukpt-ui
The intention is that these functions should indicate an error by
logging the details and then returning an empty result, and indicate
success by logging nothing and returning the populated result. The
caller will then log the result appropriately.

Note that when all of the fields required for PIN encryption/decryption
are not present, the action will be skipped without indicating an error.
This is because PIN and data actions are triggered by the same button
and if there is no PIN action to perform it will nonetheless proceed to
attempt the data action.