Implement MAC helper functions for dukpt-ui
The intention is that these functions should indicate an error by
logging the details and then returning an empty result, and indicate
success by logging nothing and returning the populated result. The
caller will then log the result appropriately.

Note that the MAC action is possible even if no input data is available.
The only requirement is that the working key type applies to the
requested MAC action.