Implement BUILD_DUKPT_UI option for CMake build
BUILD_DUKPT_UI is OFF by default and determines whether the dependencies
for dukpt-ui are mandatory, thus ensuring that it will be built. This
makes it more explicit whether dukpt-ui should be built to ensure early
failure for Github Actions that can't find Qt, instead of succeeding
without error when Qt wasn't found.