Install dukpt-tool into MacOS bundle
This will install dukpt-tool into the MacOS bundle using the bundle
content path generated for dukpt-ui such that the resulting path is
dukpt.app/Contents/bin. The intention is that the command line utility
can be bundled along with UI in the same way that CMake bundle contains
both the CMake UI and the CMake command line utilities.