Rename MacOS bundle to "Dukpt" and install dukpt-ui symlink
Installing a symlink with the original name of dukpt-ui to the same
directory as dukpt-tool inside the MacOS bundle follows the same pattern
as used by the CMake MacOS bundle. The intention is that the user can
symlink these tools from somewhere in their PATH (like /usr/local/bin)
such that they appear there with the same names as they would in a
development build.