Implement MacOS ad-hoc bundle signing
This is the first step towards proper MacOS bundle signing. Ad-hoc
signing doesn't use a signing key but instead only computes a digest and
sets a few security attributes. The intention of this implementation is
to test the appropriate time to invoke the codesign tool such that a
proper signing certificate can be configured in future.

Unfortunately Apple provides almost no documentation about ad-hoc
signing and the little available information is scattered across the
internet.