Use self-signed openemv.org code signing certificate for MacOS
This requires Github Actions to provide these secrets:
* OPENEMV_MACOS_CERT_BASE64
* OPENEMV_MACOS_CERT_PWD
* KEYCHAIN_PASSWORD

A new keychain is created, unlocked, used to import the code signing
certificate, and set as the default keychain. The keychain ACLs are also
set to allow the codesign tool to use this keychain without user
interaction. The keychain is also explicitly cleaned up by the Github
Action workflow.

The above functionality is encapsulated in scripts to allow easy reuse
but also to avoid the various secrets that are used directly on the
command line from appearing in the Github Actions log.