Update icons for dukpt-ui
The prepare_icons.sh script was created to prepare the icons for various
uses by the Linux and Windows builds of dukpt-ui. This script depends on
Imagemagick. Given that Imagemagick may delegate processing of SVGs to
various different backends that have different behaviour regarding the
alpha channel, it's best to use a PNG icon with the correct alpha
channel as the input icon.

An online tool was used to create the .icns for MacOS.