Implement transaction-originating algorithms for AES DUKPT
These algorithms are described in terms of pseudocode and this
implementation attempts to reflect similar abstractions. However, this
implementation also associates the intended cipher strength (expressed
as the AES key length) with the size of the DUKPT state object. This is
based on the assumption that any caller of these functions would already
have to manage the storage of the DUKPT state object and would therefore
have its length readily available instead of having to store an
additional key length field.