Use CPACK_VERBATIM_VARIABLES
Previously it was not only necessary to escape some characters, but also
to navigate bugs in older versions of CMake. However, CMake only
disables CPACK_VERBATIM_VARIABLES by default for backward compatibility,
suggesting that it might be the better approach in future. Newer
versions of CMake still have some bugs related to Unix paths for NSIS
but they are easier to navigate when CPACK_VERBATIM_VARIABLES is
enabled.