Implement CMAC generation and verification for TDES DUKPT
ANSI X9.24-1:2009 TDES DUKPT does not formally allow CMAC but it is
nonetheless supported by some implementations due to the weakness of
ANSI X9.19 Retail MAC. This implementation uses the same MAC key
provided by ANSI X9.24-1:2009 but instead applies it to compute CMAC.
The primary motivation for using CMAC for TDES DUKPT is when strong
authenticity and integrity is required without introducing risk to the
MAC key (as ANSI X9.19 Retail MAC would have). Any risk to the MAC key
is also a risk to the PIN key due to their XOR proximity.