Implement KBPK validator for dukpt-ui
The KBPK validator uses the key block format version to validate the
length of the KBPK. This will make it possible in future to use format
version D and E for the output of TDES DUKPT keys. This also avoids the
unnecessary restriction applied by the key validator used for DUKPT
keys, like not allowing triple length TDES, that don't apply to KBPKs.