Refactor Qt configuration
Unfortunatley the advice given by the Qt documentation to support older
Qt versions in parallel to Qt6 by using the NAMES parameter of the CMake
find_package() command, appears to make it impossible to control which
Qt version will be selected if multiple are available. In contrast to
what is stated in the documentation, this mechanism will not search for
Qt versions in the order given by the NAMES parameter because CMake
generates a list of candidate configuration file paths that doesn't
reflect this order. In addition, this mechanism appears to make it
impossible to ignore specific Qt versions by using the
CMAKE_DISABLE_FIND_PACKAGE_<PackageName> definition.

It is therefore necessary to build the NAMES list by checking for the
CMAKE_DISABLE_FIND_PACKAGE_<PackageName> definitions. This also allows
for sufficient control over the selected Qt version to comply with
Gentoo's Qt policy.