Fix code signing script for MacOS builds
It is unclear whether the keychain preparation script failed due to
changes by Homebrew or due to the recent update to use MacOS 13 for
release builds. Regardless, older versions of either Homebrew or MacOS
are no longer relevant and all that matters is that the keychain
preparation script works again.